+++ ber.h
@@ -13,7 +13,7 @@
 uint8_t hammingDistance(uint8_t n1, uint8_t n2);
 
 void toBinaryPrint(uint8_t a);
-void binaryCounter(uint8_t a, uint16_t& counter, uint16_t& comp);
+void binaryCounter(uint8_t a, uint64_t& counter, uint64_t& comp);
 #endif // !BER_H
